requirements.txt
django
django-phone-field
stripe
